# Conflicts:
#	src/main/resources/data.sql
#	src/main/resources/templates/sectionList.html
Added the student Report

+++ src/main/java/com/waa/dragons/mediationattendance/domain/TotalStudentAttendanceReport.java
@@ -14,6 +14,6 @@ public class TotalStudentAttendanceReport {
 
     private int totalSessionsPossible;
     private int totalSessionsAttended;
-    private int totalAttendancePercentage;
+    private double totalAttendancePercentage;
 
 }
